separated example and button into separate component files
