+++ example/src/App.js
@@ -1,13 +1,13 @@
 import React, { Component } from "react";
 
-import { Button } from "@amlane/styles";
+import { ExampleComponent, Button } from "@amlane/styles";
 
 export default class App extends Component {
   render() {
     return (
       <div>
-        {/* <ExampleComponent text="Modern React component module" /> */}
-        <Button text="Hello" style={{ color: "silver", background: "pink" }} />
+        <ExampleComponent text="Modern React component module" />
+        <Button text="Hello" style={{ color: "silver", background: "blue" }} />
       </div>
     );
   }
