+++ src/index.js
@@ -3,31 +3,5 @@ import PropTypes from "prop-types";
 
 import styles from "./styles.css";
 
-export default class ExampleComponent extends Component {
-  static propTypes = {
-    text: PropTypes.string
-  };
-
-  render() {
-    const { text } = this.props;
-
-    return <div className={styles.test}>Example Component: {text}</div>;
-  }
-}
-
-export class Button extends Component {
-  static propTypes = {
-    text: PropTypes.string,
-    style: PropTypes.object
-  };
-
-  render() {
-    const { text, style } = this.props;
-
-    return (
-      <button className={styles.btn} style={style}>
-        {text}
-      </button>
-    );
-  }
-}
+export { default as ExampleComponent } from "./components/ExampleComponent.react";
+export { default as Button } from "./components/Button.react";
