+++ src/styles.css
@@ -19,5 +19,5 @@
 }
 
 .btn:hover {
-  filter: brightness(95%);
+  filter: brightness(110%);
 }
